#24 Move module (series → pandas.series)
